Operation moved to a separate class. Additional check added to ShopServiceImpl
