+++ src/main/java/core/basesyntax/model/FruitTransaction.java
@@ -1,7 +1,7 @@
 package core.basesyntax.model;
 
 public class FruitTransaction {
-    private Operation operation;
+    private final Operation operation;
     private final String fruit;
     private final int quantity;
 
@@ -21,22 +21,5 @@ public class FruitTransaction {
 
     public int getQuantity() {
         return quantity;
-    }
-
-    public enum Operation {
-        BALANCE("b"),
-        SUPPLY("s"),
-        PURCHASE("p"),
-        RETURN("r");
-
-        private final String code;
-
-        Operation(String code) {
-            this.code = code;
-        }
-
-        public String getCode() {
-            return code;
-        }
     }
 }
