+++ src/main/java/core/basesyntax/strategy/OperationStrategy.java
@@ -1,16 +1,16 @@
 package core.basesyntax.strategy;
 
-import core.basesyntax.model.FruitTransaction;
+import core.basesyntax.model.Operation;
 import java.util.Map;
 
 public class OperationStrategy {
-    private final Map<FruitTransaction.Operation, OperationHandler> operationHandlers;
+    private final Map<Operation, OperationHandler> operationHandlers;
 
-    public OperationStrategy(Map<FruitTransaction.Operation, OperationHandler> operationHandlers) {
+    public OperationStrategy(Map<Operation, OperationHandler> operationHandlers) {
         this.operationHandlers = operationHandlers;
     }
 
-    public OperationHandler get(FruitTransaction.Operation operation) {
+    public OperationHandler get(Operation operation) {
         OperationHandler handler = operationHandlers.get(operation);
         if (handler == null) {
             throw new RuntimeException("Unknown type of operation: " + operation);
